misc/HtmlTestSuite/results/test-0420.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Test: Hit testing on element previously hidden by an anonymous scroll box</title>
<link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1433591" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
html, body {
  margin: 0;
  padding: 0;
}
/*
  Create a hidden scrollbox that occupies the whole viewport, then give it
  visibility: hidden dynamically. The link previously under the scrollbox
  should be clickable.
 */
.scrollable {
  position: absolute;
  top: 0;
  left: 0;
  width: 100vw;
  height: 100vh;
  overflow: scroll;
}

.scrollable .inner {
  display: table-cell;
  width: 100vw;
  height: 100vh;
}
</style>
</head><body><div class="scrollable">
  <div class="inner"/>
</div>
<a href="#">Should be clickable</a>
<script>
test(function() {
  assert_equals(document.elementFromPoint(10, 10).tagName, "DIV",
                "Should hit the scrollbox contents");
  document.querySelector('div').style.visibility = "hidden";
  assert_equals(document.elementFromPoint(10, 10), document.querySelector('a'));
}, "Link should be clickable after hiding a scrollbox with an anonymous table inside");
</script>
</body></html>
